[#711] Support encryption mode. Truncate issue still exist while rollback

+++ platform/bb/Hsqldb/src/com/rho/db/FileUtilBB.java
@@ -121,7 +121,8 @@ public class FileUtilBB implements FileAccess {
     		//if ( fc.isDirectory() )
     		//	deleteFilesInFolder(fc);
         	
-    		fconn.delete();
+    		if ( fconn != null && fconn.exists() )
+    			fconn.delete();
     	}catch(IOException exc){
     		System.out.println("FileUtilBB:delete '" + filename + "' Exception: " + exc.getMessage());
     	}finally{
